distsql: move MockDistSQLServer to distsqlpb
To avoid the distsqlrun->colrpc->distsqlrun import cycle due to colrpc
using MockDistSQLServer in tests.

Release note: None
